packages rename + FileWrite.CreateNewFileName extracted as decorator -> new class NameCreator
