+++ src/main/java/ua/javarush/encoder/cryptology/EncryptorMode.java
@@ -1,8 +1,8 @@
-package ua.javarush.encoder.Cryptology;
+package ua.javarush.encoder.cryptology;
 
 public enum EncryptorMode {
     ENCRYPT("e"), DECRYPT("d"), BRUTE_FORCE("b");
-    String abbreviation;
+    final String abbreviation;
 
     EncryptorMode(String abbreviation) {
         this.abbreviation = abbreviation;
