+++ src/main/java/ua/javarush/encoder/languageservice/DefaultLanguages.java
@@ -1,4 +1,4 @@
-package ua.javarush.encoder.LanguageService;
+package ua.javarush.encoder.languageservice;
 
 public enum DefaultLanguages {
     UKRAINIAN,ENGLISH,SYMBOLS,FRENCH
